[screenshot 2 mod, 0 new,  dark-1200px] travisdowns: make the dark mode javascript resilient
localStorage may not be available on old browsers or if the user has
disabled it, but we can't leave an unstyled page in that case!

So we fall back to the media query if LS fails, and also a final fallback
to light mode if there is an unexpected exception.